+++ WorklogManagement.UI/Components/Pages/Tracking/TrackingViewModel.cs
@@ -150,13 +150,18 @@ public class TrackingViewModel(
             return false;
         }
 
-        Worklogs = [.. Worklogs.Select(x => x.Id == savedWorklog.Id ? savedWorklog : x)];
+        Worklogs = [.. Worklogs.Where(x => x.Id != worklog.Id).Append(savedWorklog)];
 
         return true;
     }
 
     public async Task<bool> DeleteWorklogAsync(Worklog worklog)
     {
+        if (!(await _popupService.Confim("Arbeitsaufwand löschen", "Möchtest du den Arbeitsaufwand wirklich löschen?")))
+        {
+            return false;
+        }
+
         try
         {
             await _dataService.DeleteWorklogAsync(worklog.Id);
